refactor : 영화 리스트 어댑터 이름 수정, 영화 리스트 객체로 분리
